Tried to implement VoteController.cs
I tried implementing the Post method in VoteController.cs and created Vote pages.

At roollc.com/Rankings/Vote, there should be two buttons. If you click a button, the text from the buttons should show up on the screen, and the text on the buttons should change to different US states.

I am trying to do this through Vote.cshtml.cs because I want to figure out how to run the Post method and use the output in Vote.js.

Currently, the error is that nothing prints onscreen when the buttons are clicked, and the console says "Failed to load resource: the server responded with a status of 500 () [https://localhost:44364/Rankings/Controllers/Vote]".